fix: correctly place close-paren in an expression assignment containing a soak (#917)
Fixes #916

This feels a little hacky at the moment since there isn't a clear answer to when
we should insert code at the end of an expression normally vs using
`appendDeferredSuffix`. But generally I think it's reasonable to say that
`appendDeferredSuffix` should be used in cases like close-parens where we want
it to wrap the whole resulting expression (after any other deferred code).