fix: implement strictly correct argument and array spreads (#921)
Fixes #918

JS requires argument spreads to be iterable and not just array-like (even though
babel allows them to be array-like), so like with a bunch of other cases we wrap
in `Array.from` to be 100% correct. It turns out single-argument function call
spreads get compiled differently (the value doesn't get passed to `slice`), so
we need to add an additional check there to accept null or undefined.

I also needed to set up a system to distinguish assignee expressions from
non-assignee expressions. Technically, decaffeinate-parser could do this, and
attempts to in one simple case, but it seems easier and more useful to implement
it on the decaffeinate side.

Note that this is still technically not quite right for single-argument function
call spreads, since in that case non-null non-undefined non-object values (e.g.
numbers) will crash in CoffeeScript and be accepted with this defensive code.
However, that's really obscure and hopefully never comes up, and crashing less
than CoffeeScript does seems somewhat acceptable.