fix: overhaul complex expression-style classes (#922)
Fixes #835

Rather than the previous `__initClass__` wrapper function that called the
`initClass` static method, we now use an IIFE whenever the class is "complex
enough". Complex enough means one of the following:
* The class is used in an expression context (or, we can't prove it's in a
  statement context).
* The class has variables that need to be extracted.

When we do this IIFE wrapping, we also need to indent the whole class. Due to
magic-string subtleties, we need to indent the whole class body before patching
any children, so a lot of the logic needed to be refactored so we can determine
at the start whether that indentation will be necessary.

Rather than handling anonymous classes (and other similarly-tricky classes) by
passing them into a function, we generate a temporary name for them and call
`initClass` using that name. That means that a complex anonymous class in a
statement context does NOT need an IIFE.