fix: handle heregexes with a double-backslash followed by a space
Fixes #923

The lexer already breaks up the heregex into tokens of actual string content, so
we should just do escape operations over those tokens, ignoring padding
characters. Also, the escape function was looking past the end of the string,
but if `\` is at the end of a content token then we don't want to match
characters past it, so limit the skip matching to the end of the token being
considered.