fix: properly implicitly return for expression-style `if`, `switch`, and `try`
Fixes #925

In all three of these cases, we would mark our children but not ourselves as
implicitly returning. This makes sense in most cases because implicitly
returning means we get turned into an expression, which often isn't ideal. The
problem is that if we end up as an expression for any other reason (e.g. the
node is surrounded by parens), the implicit return will be skipped and the code
will be incorrect. So now we always record ourselves as implicitly returning
(which happens before we know whether we are going to be an expression), and
skip the implicit return if we end up as a statement, since we know that in that
case it's safe.