fix: properly handle reordering when patching postfix conditionals (#938)
Fixes #933

Similar to #620, the robust way to switch the order of two code snippets is to
move the right to the left, to avoid issues where an `appendLeft` at the start
of an expression doesn't get picked up by a `slice`. Rewriting postfix
conditional code to use this strategy avoids a case where we reordered a postfix
conditional around a postfix loop in a way that resulted in incorrect code.