fix: be more careful when removing whitespace to fix indentation
Fixes #935

In some cases, a line gets indented in the normalize stage, and then we try to
fix indentation by removing some indentation. This can cause too much whitespace
to be removed overall. Instead, we now check every character before removing it
to make sure it's the original single whitespace character, and skip it if not.

There are some more clever strategies that we could use if there's ever a case
where we need to remove more whitespace than we can with this algorithm, but we
crash if we can't find enough whitespace, so that should be easy to detect and
seems unlikely anyway.