fix: don't place close-parens past the end of the statement (#942)
Fixes #727

We typically prefer to put an implicit CALL_END just before to the following
close-paren if there is one, but in some cases that can cause a statement to
extend on another line before its indentation level. To avoid that, we can limit
the placement to the end of the statement that it's in by walking up the AST
until we find a BlockPatcher and limiting the insert to that statement's bounds.