fix: prefer to remove parens when a statement is surrounded by parens
Fixes #945

In CoffeeScript, regular statements, like `return`, `break`, etc, are allowed to
be surrounded by parens, and CoffeeScript emits code with the parens removed, so
we need to follow that behavior. In general, when a statement is surrounded by
parens, we might as well remove the parens and treat it as statement (if it was
going to be a statement already), since that usually looks better anyway.

This also means we can wrap statements in parens when generating them in the
normalize stage without making the code uglier, which will be nice for a later
fix.