fix: allow sequence operations to be patched as statements (#949)
Fixes #941

Previously, SeqOp was a binary operator, so it forced both of its operands to be
expressions. But CoffeeScript is actually a little looser and allows either one
to be a statement if the SeqOp itself is a statement. This is why you can
sometimes use `break`, `continue`, or `return` inside a SeqOp. To fix, I change
it to no longer be a binary operator and only force its children to be
expressions when it is an expression.