feat: handle soaked new operations (#951)
Fixes #383

This ended up being pretty much the same logic as soaked function application,
except I needed to tweak the prefix code so `new` is part of the prefix. Also,
since `new` always calls the function with a fresh `this`, we can skip some of
the more complex cases in soaked function application that were ensuring that
the proper `this` was set.

This also pulls in a decaffeinate-parser version with a SoakedSlice node that's
not handled currently, but those always crashed anyway, so I'll handle those in
a later commit.