refactor: get rid of source map code
Closes #119

It didn't work anyway, since esnext doesn't support source maps, and made it
sort of a pain to add new stages or work with the code outside of stages.

I still kept around the response format of an object with a `code` property
rather than making it just a string, since I have vague plans for decaffeinate
outputting a list of warnings or cleanup suggestions or things like that, and
having a flexible output format would make that easier.