fix: improve close-brace placement for arrow functions (#955)
Fixes #787

We already had some custom logic for non-arrow functions to place the
close-brace right before a close-paren on the next line, but that wasn't
happening for arrow functions. To fix, I factored it out into shared code used
in both places, and tweaked some of the logic in case the close-paren is at the
end of the line.