fix: be more careful about implicit close-paren placement
Fixes #963

When placing an implicit close paren for a function call, in some cases, it's
fine to place it either at the end of the line or before the close-paren on the
next line, in some cases, but in some cases one or the other is required. From
what I can tell, for normal expressions, it should be placed at the end of the
line, but if the last arg has a function, `if`, or other block, the close-paren
may need to be on the next line for the block to be properly closed. There are
likely more nuances here, but this change should at least handle more cases.