fix: don't remove parens for implicitly-returned expressions
Fixes #978

The new spec for implicit returning is that the parens are always kept and must
be handled by the implicit return code if necessary. For `return`, we want to
keep the parens to avoid ASI issues (and since we know it will always be an
expression). For `break` in `switch`, we need to remove the parens if they
exist, so add a special case for that. In loops, the issue doesn't seem to come
up; parens aren't allowed around `break` or `continue` in loops.