fix: properly set canPatchAsExpression for TryPatcher
Fixes #986

In some cases, we would try using `map` instead of an IIFE for a try within a
loop, which only works if the try can actually be used in an expression context.
When a try has a statement like `continue`, we need to disallow it from becoming
an expression.