fix: skip implicitly-returned break and continue in more cases
Fixes #987

Since we can generally assume the CoffeeScript we get is valid, it's generally
the case that any `break` and `continue` statements that are implicitly returned
should just be kept as-is, so we now apply that more broadly in `NodePatcher`.
In particular, a `break` within a non-IIFE switch that's implicitly returned was
causing a crash.