Change canPatchAsExpression to be more precise
Fixes #993

`throw` is an example of a node type that can patch as an expression, but
prefers not to. Our existing BlockPatcher code was using that to say that the
block could not be patched as an expression when actually it could, so
decaffeinate was crashing on valid code. Fixing this had some cascading effects,
and generally some more code needed to be more precise about prefering an
expression vs requiring an expression.