fix: properly handle negated parenthesized soak operations
Fixes #995

When a node is negated and then changed to a ternary as part of a soak
operation, we weren't always adding parens when we needed to. In particular,
when the expression was already surrounded in parens, the soak code would view
extra parens as unnecessary even though a `!` had been inserted within those
parens. Now, we track when we negate in that way and always add parens in that
case when changing a soak container to a ternary.