feat: add support for Literate CoffeeScript
Fixes #714

I maybe should have taken a simpler approach, but this was a fun exercise.
There's now an optional stage at the start that turns literate CoffeeScript into
regular CoffeeScript with some tweaks to make the result look nice. I don't use
magic-string, and instead just produce an array of lines that gets joined into
the final source code, which seems fine because source maps are unlikely to
happen and going line-by-line is a little easier for this case.