fix: avoid block comments when processing literate coffeescript
Since CoffeeScript block comments have a meaningful indentation level, it's
actually unsafe to insert them anywhere in the file, since indentation
differences can cause the normalize stage to break. Instead, we now take the
boring approach of always using `//` comments. Theoretically, we could be
smarter about this and convert them to multiline comments on the JS side where
the indentation doesn't matter, but that seems like it's probably not worth it.