fix: always replace the extension with .js (#1008)
Fixes #1007

This fixes a regression caused by #1005 where the computed filename was only
changed if the original filename ended in .coffee, .litcoffee or .coffee.md.
Really, we want to keep the old behavior and always change whatever file
extension we find (or just add .js if there's no extension). But to properly
handle .coffee.md we should just special-case that.

I tested this manually, but plan to add automated tests for the CLI pretty soon.