feat: add a --modernize-js option to modify existing JS files (#1018)
This is useful for files that have already been converted but could use some
additional transformations, like converting `require` to `import`. It mostly
just calls through to esnext, but is nice in that it uses the decaffeinate
options, so it should be easier to hook things up to bulk-decaffeinate with this
approach.

To make this work, I had to modify the code so that when converting files on
the filesystem, we read the file as a whole convert the code, then write the
file. Having a read and write stream open for the same file at the same time was
causing trouble.