fix: properly handle \0 chars in heregexes
Fixes #1028

In other cases, we don't remove intermediate whitespace, so we don't need to
worry, but for heregexes we might e.g. have `\0 1`, which would convert to
`\01`, which is an octal literal. Instead, we always convert `\0` to `\x00`.
This is a little dumber than CoffeeScript's approach, since we always do the
conversion in heregexes and CoffeeScript only does it if the next character is
0-7, but this case seems rare enough that we don't need to be too clever about
it.