test: rework the validate function to be more flexible (#1046)
Previously it was doing a JSON conversion to avoid the issue of prototype
differences between V8 contexts, but now it uses a custom deep equality function
instead.

Also, it now compares the result when run directly in V8, to guard against
differences between babel and V8. This required changing the spec to call a
function on the result rather than assigning the result to a variable.