fix: preserve order for `in` operations when necessary
Fixes #1038

This makes some cases uglier, unfortunately, but from a scan through my codebase
it seems relatively rare to have impure operands to the `in` operator.

I also added an `isPure` method that is very similar to `isRepeatable`, but
doesn't interfere with some cases where it's better for an expression to not be
repeatable even if it's technically pure.