fix: add a custom code path for the extracted variable `in` case
Fixes #1052

Since the full expressions are in order for this case, it's best to just patch
in-place rather than getting the code, since otherwise some inserted
magic-string fragments can be left over and cause issues. Since the patching
strategy is different enough, I just made it a completely separate function.