fix: properly handle bound static methods (#1060)
Fixes #1056

The existing `initClass` code path already properly handled this case, since the
arrow function binds the proper `this`, but we were treating it as a regular
method and effectively ignoring the fact that it was bound. To fix, we can just
add a special case to move it to `initClass`.