fix: extract loop variables to their own name when necessary
Fixes #1044

We need to do this when the loop body might assign to the loop variable. We can
detect this by determining if there are any assignments to the variable directly
in the loop body or if there are any instances of the variabble assigned from a
closure anywhere in the function scope.