fix: make super code more defensive about computed names
Fixes #1082

In both computed method names and when computing the class to use in a dynamic
method assignment, we now save it to a variable so that at `super` time we know
it hasn't changed.

This makes some existing code uglier, and it could probably be improved by
adding some additional checks to see if the variables are ever reassigned or
modified, but all of these cases are obscure enough that this is probably fine
for now.