fix: rework loop index saving to happen within the init and update
Fixes #1083

In order to have the proper index value both on the first iteration and after
the loop is finished, we need to put the index saving code in both the loop init
and the loop update code.

This also ends up implementing index saving for all types of loops, I think;
previously, range loops had the same bug.

The scope-related details are the same, but the other details have been
reworked; rather than there being an "index binding" and a "user-specified index
binding", we now have an "internal index binding" and an "index binding". The
loop details work with the internal binding, and save to the normal index
binding if they're different.