fix: properly handle super calls within static methods in initClass (#1093)
Fixes #860

This `initClass` transform can reposition static methods so that they no longer
have the proper `super` semantics according to CoffeeScript's algorithm.
Instead, the `super` matches the `initClass` method itself. Since the code after
the normalize stage is incorrect, there's no way to fix this properly in the
main stage; instead, this commit moves some of the `super` logic to the
normalize stage, while most of it is still in the main stage.

This ended up being kind of a pain, but seems to be working. Some details:
* One role of the normalize stage is to add explicit arguments to all bare super
  calls. Since I'm now doing some super transformations in the normalize stage,
  I need to handle those two cases specifically.
* We need to extract out the `this` reference, but an expression like `@a` would
  then become `(cls = @)a`, which doesn't parse. To address this, we now have a
  special case for patching that form of `this` as repeatable in the normalize
  stage.
* The `containsSuperCall` method needed to be extracted out into a shared
  function so it's usable from the normalize stage.
* We need to do some custom logic to detect if the super call should be
  early-transformed in the first place. It needs to be a static method
  declaration within a class, but not a top-level one that would become a
  regular static method.
* The code snippet for manually calling super for a static method is different
  than for an instance method, so the normalize case uses the static method code
  snippet and the main stage case uses the instance method code snippet.