fix: properly add variable declarations for IIFE loop assignees (#1094)
Fixes #750

The solution here is similar in spirit to #639, but has some differences.
Rather than allowing the loop to become an IIFE and letting
add-variable-declarations put in the declaration for us, we now explicitly add
declarations for any index/value bindings that we know will be in IIFEs, as long
as there is at least one usage of the variable outside of the loop. This
required a lot of coordination between BlockPatcher and all of its descendant
ForPatchers during initialize so that patching can be done in the proper order.