fix: remove semicolons when turning blocks into expressions (#1109)
Fixes #1108

Now that all lone semicolons are treated in the AST as empty blocks, they turn
into `undefined` as expected when treated as an expression, but the semicolon is
still there. To fix this, we can just detect if the block ends in a semicolon
(even for non-empty blocks) and remove it.