refactor: add an initial suggestions system that isn't hooked up
Progress toward #1107

At the moment, we never print suggestions, but we do return them from the
convert function. Each suggestion has a shorthand code and a summary string, and
doesn't include any context like where in the code the transformation happened.
Suggestions are sorted by code, although in the future it may be better to
independently sort them by priority.