feat: insert suggestion comments in generated code
Progress toward #1107

Rather than printing suggestions out to the console, I think it actually makes
more sense to include them in the code. This should make them really hard to
miss and be especially useful for the use case of doing a large conversion and
then going through each file as a follow-up task. I added a `disableSuggestions`
flag to disable this behavior if people want.

For now, the `disableSuggestions` is always set to true from the CLI, meaning
the feature can't be accessed normally now; it's just exposed to tests. I'll
change it to false by default soon along with all of the other breaking changes.