feat: enable suggestions system, update default options (#1121)
Fixes #1107

Implement several defaults changes as described in #1107. This involved updating
the options structure, the CLI code, the docs, and the tests. This also includes
various other little cleanups to the docs and generated code to better match the
new system.

Also rename `disableSuggestions` to `disableSuggestionComment`, since maybe in
the future there will be a different way of displaying suggestions.

BREAKING CHANGE:

JavaScript files produced by decaffeinate now have a comment at the top of the
file with suggestions on cleanup steps. The new `--disable-suggestion-comment`
flag disables this behavior.

decaffeinate now prefers `const` for variables when possible. The
`--prefer-const` option is now a no-op, and a new `--prefer-let` option opts
into the old behavior.

decaffeinate no longer tries to convert code to JS modules by default. There are
three configurations to choose from:
* To keep code as CommonJS (or whatever the original code was using), specify no
  options. Previously, this was the `--keep-commonjs` option.
* To convert code to JS modules in a way that always generates a default export
  for correctness reasons, specify `--use-js-modules`. Previously, this was the
  `--force-default-export` option.
* To convert code to JS modules and use named exports on a best-effort basis,
  specify `--use-js-modules --loose-js-modules`. Previously, this was the
  default option.

decaffeinate now automatically inserts a code snippet at the top of constructors
using `this` before `super`, which allows them to behave as expected, but only
when using Babel or TypeScript under certain configurations. There are three
configurations to choose from:
* To insert the Babel/TypeScript workaround code snippet when necessary, specify
  no options. Previously, this was the `--enable-babel-constructor-workaround`
  option.
* To produce code that uses `this` before `super` (which is well-formed but
  crashes at runtime), specify `--disable-babel-constructor-workaround`.
  Previously, this was the `--allow-invalid-constructors` option.
* To make decaffeinate give an error when code would use `this` before `super`,
  specify `--disallow-invalid-constructors`. Previously, this was the default
  behavior.