refactor: rework parents and scopes to be stored separately
Previously, the scope and the parent were tacked onto AST nodes, but this
doesn't work particularly well with a type system. Instead, we now compute a
`DecaffeinateContext` that is used where `ParseContext` was previously used, and
also provides the scope and parent of each node. This replaces the `parse`
function, and I had to move `Scope` to typescript as part of this change as
well. I also changed the `traverse` implementation to provide parents, and got
rid of some dead code.

Probably it would be better if decaffeinate-parser computed the parents, but at
the moment it's a little easier on the decaffeinate side, and I need to store
scopes separately anyway.