feat: add an option to convert soak operations to JS optional chaining (#1158)
Progress toward #336

This implements soaked member access, dynamic member access, function
invocation, and slice/splice using the JS optional chaining feature.

One difference between the CS and JS syntax that this diff accounts for is that
JS does not let you use `?.` on variable that hasn't been defined. We simply
fall back to the normal code path in that case, which does a typeof check.

There are two known issues that will be fixed in follow-up PRs:
* In CS, a soaked function call checks `typeof callee === 'function'`, whereas
  in JS it checks against null or undefined. One idea is to define a wrapper
  helper that converts non-functions to null or undefined, but a simple
  implementation of that was causing issues with the function call using the
  right `this`.
* In CS, parentheses stop the soak from propagating, so `(a?.b).c` is different
  from `a?.b.c`. Once idea is to change it to an equivalent expression, like
  `(0, a?.b).c`, as a way of forcibly stopping the soak operation.

Both of these issues are relatively obscure, and might change semantics in the
proposal anyway.