fix: allow expanding the bounds of parenthesized function bodies
Fixes #1172

Ultimately, the issue was we had a block like `(a)` where `isSurroundedByParens`
was returning true. This meant that the bounding patcher was itself, even
though the block itself wasn't really surrounded by parens, it just started and
ended with parens. To fix, we can say the node isn't surrounded by parens in
this case, which means we're allowed to insert contents before the start of the
block.