fix: handle double commas in arg lists by removing the first
Fixes #1182

An implicit object literal followed by an arg separator can sometimes result in
two commas in a row between arguments. The previous strategy was to just remove
the second, but in some cases, that could cause indentation issues that change
the parsing so there's no more argument separator. Removing the first one
instead should fix this problem.

Another fix may be to fix the location data of implicit object literals to
include the trailing comma, but this fix should work for now.