fix: don't normalize commas to semicolons outside of block bounds
Fixes #1195

The fix for #1191 relied on a change that had us editing beyond the block
bounds, which was causing lots of other problems. For now, I'll just limit the
comma-to-semicolon conversion to instances within the block, which I think
should fix the regressions but re-breaks #1191. The more fundamental issue in #1191
is bad parser location data (the comma isn't being included in any node bounds),
so probably the better fix will be in decaffeinate-parser or
decaffeinate-coffeescript.