fix: don't generate blockless arrow functions that contain assignments
Fixes #1202

add-variable-declarations won't have a place to add a declaration if the
function doesn't have a block body, so if there's any assignment in an arrow
function, we always give it a block body.

The formatting after decaffeinate and add-variable-declarations looks really
bad, I think because add-variable-declarations doesn't attempt to handle inline
functions, but this case is rare enough that I'm not too worried about
formatting.

Also update add-variable-declarations and add a regression test for #1201.