fix: change continuation patching to happen before program patching
Fixes #1230

Generally with magic-string it's best to patch in order, and to remove before
you add. In this case, we were removing continuation tokens after the fact
across the whole program, which was causing an open-paren to be incorrectly
removed since it was in the range being removed. Removing continuations first
seems to avoid this problem, and also avoid the need to wrap the remove in a
try/catch.