feat: add a `--use-cs2` option and implement LHS array spread
Progress toward https://github.com/decaffeinate/decaffeinate/issues/427

There's still new syntax and semantics to support, but this gets the beginnings
of a CS2 mode.

For array spread on the left-hand side, we now detect where the ellipsis is and
don't change it if it's already on the left. We also skip Array.from in CS2
mode.