fix: remove parens from the inside of a block instead of the outside
The parser details for blocks surrounded by parens have been tweaked so that the
block bounds include the parens. That means we can't just remove parens
surrounding the block; we need to find interior parens not attached to a child
and remove those. This fixes a test regression from a decaffeinate-parser
change.