fix Babel/TypeScript `super` hack to work w/ istambul
we use istambul in our tests to calculate code coverage for our app.
when it runs it wraps each statement with some stuff to track what lines
were executed. when it does that it adds an extra semicolon to:
`let thisFn = (() => { return this; })`
so this change it to have it just look for the last ";"
(the one right after `this`)