feat: add optional-chaining
This turned out to be not that straightforward. You can't simply leave the CoffeeScript code alone since in some cases the JavaScript equivalent has an extra `.`. Still, that part was pretty easy. The harder part is that CoffeeScript soaked expressions can be used in place that JavaScript optional chaining cannot, such as in an assignee: `a?.b = 1` is valid CoffeeScript, but not JavaScript. To address this I added `isInsideAssignee` to check whether a node should be considered part of an assignee or not, and to bail early if it is. Otherwise you'd end up with a parse error in a later stage.