fix(class): ensure multiple bindings work in generated constructor
Annoyingly, TypeScript didn't catch this because adding non-string things to a string is allowed since they get coerced to a string. For arrays, it calls `join()` on the array which by default joins with a comma. With a single binding this happened to work, but it breaks with multiple bindings.

Closes #2169