feat(cli): allow passing in IO streams
Rather than hardcoding `process.stdin` etc, this allows passing in the IO streams to use. We then use this in the CLI tests to make them faster and stop requiring a full build before running them.

BREAKING CHANGE: The `args` passed to `run` must now be the full `process.argv`, including the path to NodeJS and the script path.