Bump pnpm and Lockfile Version
Switch from pnpm7 to pnpm8, seeing as pnpm9 is the most recent
release. This is accurate to what versions of pnpm are fresh in
package managers other than npm, such as nixpkgs.

Switch from lockfile v5.4 to v6.0 for compatibility with pnpm8.
v5 is recently deprecated, and causes an error when freezing the
lockfile for pnpm8 and above.